// front/src/api/generated/models/AsyncSubmitResp.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AsyncSubmitResp = {
    /**
     * 提交 ID，用于查询提交结果
     */
    submitId: string;
    /**
     * 显示给用户的提示
     */
    tip?: string;
};