// front/src/api/generated/models/IntroVideo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 课堂导入视频
 */
export type IntroVideo = {
    /**
     * 是否需要自动播放导入视频
     */
    autoPlay: boolean;
    /**
     * 视频的 URL (支持相对或绝对链接)
     */
    url?: string | null;
};